+++ src/components/NavBar/assets/gekko-logo.png
