+++ src/utils/truncateText.ts
@@ -1,4 +1,4 @@
 export const truncateText = (str: string) => {
-    if(str.length < 25) return str;
+    if(str.length < 40) return str;
     return str.substring(0, 25) + '...';
 }
