--- utils/formatPrice.ts
@@ -1,7 +1,0 @@
-export const formatPrice = (amount: number) => {
-    return new Intl.NumberFormat("ru",
-    {
-        style: "currency",
-        currency: "RUB"
-    }).format(amount);
-}
